Added user_weight field on Room database and API response

+++ app/src/main/java/br/net/easify/tracker/helpers/TrackerHelper.kt
@@ -35,7 +35,7 @@ class TrackerHelper {
             return convertMetersPerSecondsIntoKilometerPerHour(averageSpeedInMetersPerSeconds)
         }
 
-        fun calculateCalories(weigth: Double, path: List<DbRoutePath>): Double {
+        fun calculateCalories(weigth: Int, path: List<DbRoutePath>): Double {
 
             val speed = calculateAverageSpeedInKmPerHour(path)
